Updated token and crowdsale design Added reserved tokens functionality Added distribution and verification functionality to transfer tokens after ICO finish

+++ contracts/StagedCrowdsale.sol
@@ -5,19 +5,19 @@ import './SafeMath.sol';
 
 contract StagedCrowdsale is Ownable {
 
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
     struct Stage {
-        uint hardcap;
-        uint price;
-        uint invested;
-        uint closed;
+        uint256 hardcap;
+        uint256 price;
+        uint256 invested;
+        uint256 closed;
     }
 
     Stage[] public stages;
 
-    function getCurrentStage() public constant returns(uint) {
-        for(uint i=0; i < stages.length; i++) {
+    function getCurrentStage() public constant returns(uint256) {
+        for(uint256 i=0; i < stages.length; i++) {
             if(stages[i].closed == 0) {
                 return i;
             }
@@ -25,7 +25,7 @@ contract StagedCrowdsale is Ownable {
         revert();
     }
 
-    function addStage(uint hardcap, uint price) public onlyOwner {
+    function addStage(uint256 hardcap, uint256 price) public onlyOwner {
         require(hardcap > 0 && price > 0);
         Stage memory stage = Stage(hardcap.mul(1 ether), price, 0, 0);
         stages.push(stage);
